chore: Rename provders to appProviders
